backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
moviepy==1.0.3
numpy==2.1.1
openapi-codec==1.3.2
pillow==10.4.0
proglog==0.1.10
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==75.1.0
simplejson==3.19.3
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
